Split Auth and JWT Guard tests into its own file
